survey into compt and data into a static folder
