+++ my-app/src/components/Survey/MentalActivitiesStep.tsx
@@ -1,0 +1,39 @@
+import React from "react";
+import { mentalActivities } from "../../data/static/surveyData";
+
+interface MentalActivitiesStepProps {
+  data: any;
+  onChange: (field: string, value: any) => void;
+}
+
+const MentalActivitiesStep: React.FC<MentalActivitiesStepProps> = ({
+  data,
+  onChange,
+}) => {
+  return (
+    <div className="survey__step">
+      <h2 className="survey__step-title">Mental & Learning Activities</h2>
+      <p className="survey__step-subtitle">
+        Which activities do you enjoy? Select all that apply
+      </p>
+
+      <div className="survey__activities-grid">
+        {mentalActivities.map((activity) => (
+          <div key={activity.key} className="survey__activity-item">
+            <label className="survey__activity-checkbox">
+              <input
+                type="checkbox"
+                checked={data[activity.key] || false}
+                onChange={(e) => onChange(activity.key, e.target.checked)}
+              />
+              <span className="survey__activity-icon">{activity.icon}</span>
+              <span className="survey__activity-name">{activity.name}</span>
+            </label>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default MentalActivitiesStep;
