+++ my-app/src/components/Survey/survey.scss
@@ -1,21 +1,40 @@
 .survey {
   min-height: 100vh;
   background: linear-gradient(135deg, #fefefe 0%, #f8f6f3 100%);
-  padding: 2rem 0;
+  padding: 6rem 0 2rem 0;
 
   &__container {
     max-width: 800px;
     margin: 0 auto;
     padding: 0 2rem;
 
+    @media (max-width: 1024px) {
+      max-width: 90%;
+      padding: 0 1.5rem;
+    }
+
     @media (max-width: 768px) {
+      max-width: 95%;
       padding: 0 1rem;
+    }
+
+    @media (max-width: 480px) {
+      max-width: 100%;
+      padding: 0 0.75rem;
     }
   }
 
   &__header {
     text-align: center;
     margin-bottom: 3rem;
+
+    @media (max-width: 768px) {
+      margin-bottom: 2rem;
+    }
+
+    @media (max-width: 480px) {
+      margin-bottom: 1.5rem;
+    }
   }
 
   &__title {
@@ -24,8 +43,17 @@
     color: #2d3748;
     margin-bottom: 1rem;
 
+    @media (max-width: 1024px) {
+      font-size: 2.25rem;
+    }
+
     @media (max-width: 768px) {
       font-size: 2rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 1.75rem;
+      margin-bottom: 0.75rem;
     }
   }
 
@@ -33,6 +61,16 @@
     font-size: 1.125rem;
     color: #4a5568;
     margin-bottom: 2rem;
+
+    @media (max-width: 768px) {
+      font-size: 1rem;
+      margin-bottom: 1.5rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 0.875rem;
+      margin-bottom: 1rem;
+    }
   }
 
   &__progress {
@@ -69,8 +107,19 @@
     padding: 2rem;
     margin-bottom: 2rem;
 
+    @media (max-width: 1024px) {
+      padding: 1.75rem;
+    }
+
     @media (max-width: 768px) {
       padding: 1.5rem;
+      margin-bottom: 1.5rem;
+    }
+
+    @media (max-width: 480px) {
+      padding: 1rem;
+      margin-bottom: 1rem;
+      border-radius: 0.75rem;
     }
   }
 
@@ -84,11 +133,34 @@
     font-weight: 700;
     color: #2d3748;
     margin-bottom: 0.5rem;
+
+    @media (max-width: 1024px) {
+      font-size: 1.625rem;
+    }
+
+    @media (max-width: 768px) {
+      font-size: 1.5rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 1.25rem;
+      margin-bottom: 0.25rem;
+    }
   }
 
   &__step-subtitle {
     color: #4a5568;
     margin-bottom: 2rem;
+
+    @media (max-width: 768px) {
+      font-size: 0.875rem;
+      margin-bottom: 1.5rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 0.8rem;
+      margin-bottom: 1rem;
+    }
   }
 
   &__form {
@@ -107,6 +179,14 @@
     font-weight: 600;
     color: #2d3748;
     font-size: 1rem;
+
+    @media (max-width: 768px) {
+      font-size: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 0.8rem;
+    }
   }
 
   &__input,
@@ -116,6 +196,17 @@
     border-radius: 0.5rem;
     font-size: 1rem;
     transition: border-color 0.3s ease;
+
+    @media (max-width: 768px) {
+      padding: 0.625rem 0.875rem;
+      font-size: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      padding: 0.5rem 0.75rem;
+      font-size: 0.8rem;
+      border-radius: 0.375rem;
+    }
 
     &:focus {
       outline: none;
@@ -141,11 +232,30 @@
     background: rgba(212, 165, 116, 0.1);
     border: 1px solid rgba(212, 165, 116, 0.2);
     border-radius: 0.75rem;
+
+    @media (max-width: 768px) {
+      gap: 0.75rem;
+      padding: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      gap: 0.5rem;
+      padding: 0.75rem;
+      border-radius: 0.5rem;
+    }
   }
 
   &__note-icon {
     font-size: 1.5rem;
     flex-shrink: 0;
+
+    @media (max-width: 768px) {
+      font-size: 1.25rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 1rem;
+    }
   }
 
   &__note-content {
@@ -177,8 +287,19 @@
     grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
     gap: 1rem;
 
+    @media (max-width: 1024px) {
+      grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
+      gap: 0.875rem;
+    }
+
     @media (max-width: 768px) {
       grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+      gap: 0.75rem;
+    }
+
+    @media (max-width: 480px) {
+      grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+      gap: 0.5rem;
     }
   }
 
@@ -214,6 +335,17 @@
     transition: all 0.3s ease;
     border-radius: 0.75rem;
 
+    @media (max-width: 768px) {
+      gap: 0.5rem;
+      padding: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      gap: 0.375rem;
+      padding: 0.75rem;
+      border-radius: 0.5rem;
+    }
+
     &:hover {
       background: rgba(212, 165, 116, 0.05);
     }
@@ -227,6 +359,11 @@
       width: 18px;
       height: 18px;
       accent-color: #d4a574;
+
+      @media (max-width: 480px) {
+        width: 16px;
+        height: 16px;
+      }
     }
   }
 
@@ -239,6 +376,14 @@
 
   &__activity-icon {
     font-size: 2rem;
+
+    @media (max-width: 768px) {
+      font-size: 1.75rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 1.5rem;
+    }
   }
 
   &__activity-name {
@@ -246,6 +391,14 @@
     font-weight: 600;
     color: #2d3748;
     margin: 0;
+
+    @media (max-width: 768px) {
+      font-size: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 0.8rem;
+    }
   }
 
   &__activity-options {
@@ -264,6 +417,16 @@
     display: grid;
     grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
     gap: 0.75rem;
+
+    @media (max-width: 768px) {
+      grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
+      gap: 0.5rem;
+    }
+
+    @media (max-width: 480px) {
+      grid-template-columns: repeat(auto-fit, minmax(80px, 1fr));
+      gap: 0.375rem;
+    }
   }
 
   &__checkbox {
@@ -319,6 +482,24 @@
     justify-content: space-between;
     align-items: center;
     gap: 1rem;
+
+    @media (max-width: 768px) {
+      gap: 0.75rem;
+    }
+
+    @media (max-width: 480px) {
+      gap: 0.5rem;
+    }
+
+    // Ensure Next button is always on the right, even when Back button is hidden
+    .survey__btn--primary {
+      margin-left: auto;
+    }
+
+    // When Back button is present, reset the margin
+    .survey__btn--secondary + .survey__btn--primary {
+      margin-left: 0;
+    }
   }
 
   &__btn {
@@ -330,6 +511,21 @@
     transition: all 0.3s ease;
     font-size: 1rem;
 
+    @media (max-width: 1024px) {
+      padding: 0.75rem 1.75rem;
+    }
+
+    @media (max-width: 768px) {
+      padding: 0.625rem 1.5rem;
+      font-size: 0.875rem;
+    }
+
+    @media (max-width: 480px) {
+      padding: 0.5rem 1.25rem;
+      font-size: 0.8rem;
+      border-radius: 0.375rem;
+    }
+
     &--primary {
       background: linear-gradient(135deg, #d4a574, #b8860b);
       color: white;
@@ -337,6 +533,13 @@
       &:hover {
         transform: translateY(-2px);
         box-shadow: 0 4px 12px rgba(212, 165, 116, 0.3);
+      }
+
+      @media (max-width: 480px) {
+        &:hover {
+          transform: none;
+          box-shadow: none;
+        }
       }
     }
